Upgrade `rules_nodejs` to 4.7.0
We tried to just upgrade to 4.0.0 to adjust to the breaking changes first, and
then upgrade to another minor release, but no 4.x versions were working for us,
and 4.7.0 was the first and only one that worked. Incidentally, 4.7.0 also
seems like the last 4.x release before the 5.x series, so we'll have to make
another major version upgrade soon.

For historical reference, here's the command we used for testing:

```sh
$ bazel run //third_party/stimulus/hello-world-ts:devserver
```

and the errors we were getting were one of the following:

> ERROR: error loading package 'third_party/stimulus/hello-world-ts': at [...]/.cache/bazel/_bazel_[...]/61a525b3740f61615718cb5fbe184459/external/npm/@bazel/typescript/index.bzl:22:6: at [...]/.cache/bazel/_bazel_[...]/61a525b3740f61615718cb5fbe184459/external/npm/@bazel/typescript/internal/ts_project.bzl:6:6: cannot load '@build_bazel_rules_nodejs//third_party/github.com/bazelbuild/bazel-skylib:lib/partial.bzl': no such file

or

> ERROR: error loading package 'third_party/stimulus/hello-world-ts': at [...]/.cache/bazel/_bazel_[...]/61a525b3740f61615718cb5fbe184459/external/npm/@bazel/typescript/index.bzl:22:6: at [...]/.cache/bazel/_bazel_[...]/61a525b3740f61615718cb5fbe184459/external/npm/@bazel/typescript/internal/ts_project.bzl:7:6: cannot load '@build_bazel_rules_nodejs//third_party/github.com/bazelbuild/bazel-skylib:rules/build_test.bzl': no such file

One of these errors appeared for all versions of `rules_nodejs` we tested
(which was most of 4.x series) until 4.7.0, which worked.

`rules_nodejs` had a number of changes in the 4.0.0 major release:

* https://github.com/bazelbuild/rules_nodejs/releases/tag/4.0.0

To follow the default versions, we have made similar upgrades:

* Node updated to then-current LTS (14.17.5)
* Yarn updated to then-latest (1.22.11)
* updated Node to v15 in the GitHub Actions config

There isn't v4.7.0 for either `@bazel/concatjs` or `@bazel/typescript`, so they
were upgraded to the latest 4.x version, which is 4.6.2 for both of them.