Upgrade `webpack-cli` to 4.9.2
This upgrade required making a bunch of changes to keep things working.

First, we ran into this error:

> ERROR:
> [...]/.cache/bazel/_bazel_[...]/61a525b3740f61615718cb5fbe184459/external/io_bazel_rules_closure/java/com/google/javascript/jscomp/BUILD:19:13:
> error loading package '@com_google_protobuf//': cannot load
> '@bazel_skylib//rules:common_settings.bzl': no such file and referenced by
> '@io_bazel_rules_closure//java/com/google/javascript/jscomp:jscomp'

As a result, we added Bazel Skylib to `WORKSPACE`.

The next error was:

> Module not found: Error: Can't resolve
> 'bazel-out/k8-fastbuild/bin/third_party/stimulus/hello-world-ts/hello.mjs' in
> '[...]/.cache/bazel/_bazel_[...]/61a525b3740f61615718cb5fbe184459/sandbox/linux-sandbox/259/execroot/using_stimulus'
>
> Did you mean
> './bazel-out/k8-fastbuild/bin/third_party/stimulus/hello-world-ts/hello.mjs'?
>
> Requests that should resolve in the current directory need to start with
> './'.
>
> Requests that start with a name are treated as module requests and resolve
> within module directories (node_modules).
>
> If changing the source code is not an option there is also a resolve options
> called 'preferRelative' which tries to resolve these kind of requests in the
> current directory too.

This was fixed by adding `preferRelative: true` to `webpack.config.js`.

Finally, what may be a behavior change in either Bazel Closure rules or Webpack,
the output which used to be a single file `bundle.js` became a directory, with
a file `main.js` inside, so we updated it from outputting `bundle.js` to
referencing `bundle/main.js` instead which required changing both the `BUILD`
file as well as `index.html` for the sample TypeScript app.